feat: 마이페이지 data store로 관리
