+++ src/pages/mypage/ProfilePage.tsx
@@ -5,23 +5,8 @@ import ProfileImg from '../../assets/images/profile.png'
 import styled from 'styled-components'
 import axios from 'axios'
 import { useEffect, useState } from 'react'
-import { useApiUrlStore } from '../../store/store.ts'
+import { useApiUrlStore, useProfileDataStore } from '../../store/store.ts'
 
-interface ProfileData {
-  name: string
-  nickname: string
-  part: string
-  email: string
-  tel: number
-  interests: string
-  blogUrl: string
-  publicRelations: string
-  job: string
-  heart: number
-  starAverage: number
-  solved: number
-  matchingCount: number
-}
 
 const Container = styled.div`
   display: flex;
@@ -109,22 +94,7 @@ const Detail = styled.div`
 
 function ProfilePage() {
   const { apiUrl } = useApiUrlStore()
-
-  const [profileData, setProfileData] = useState<ProfileData>({
-    name: '',
-    nickname: '',
-    part: '',
-    email: '',
-    tel: 0,
-    interests: '',
-    blogUrl: '',
-    publicRelations: '',
-    job: '',
-    heart: 0,
-    starAverage: 0,
-    solved: 0,
-    matchingCount: 0,
-  })
+  const {profileData, setProfileData} = useProfileDataStore()
 
   const getProfile = async () => {
     try {
